+++ analysis/online/lisa/LisaOnlineSpectraDaq.cxx
@@ -501,8 +501,8 @@ void LisaOnlineSpectraDaq::Exec(Option_t* option)
         city = lisaCalItem.Get_city();
         int xpos = lisaCalItem.Get_xposition();
         int ypos = lisaCalItem.Get_yposition();
-        uint32_t energy = lisaCalItem.Get_energy();
-        uint32_t energy_MWD = lisaCalItem.Get_energy_MWD();
+        float energy = lisaCalItem.Get_energy();
+        float energy_MWD = lisaCalItem.Get_energy_MWD();
         trace = lisaCalItem.Get_trace_febex();
         int pileup = lisaCalItem.Get_pileup();
         int overflow = lisaCalItem.Get_overflow();
